Move GET /datasets migration tests to the endpoint test file
