+++ tests/routers/openml/migration/datasets_migration_test.py
@@ -1,153 +1,12 @@
 import asyncio
-import json
 import re
 from http import HTTPStatus
 
 import httpx
 import pytest
 
-import tests.constants
 from core.conversions import nested_remove_single_element_list
 from tests.users import ApiKey
-
-
-@pytest.mark.parametrize(
-    "dataset_id",
-    range(1, 132),
-)
-async def test_dataset_response_is_identical(  # noqa: C901, PLR0912
-    dataset_id: int,
-    py_api: httpx.AsyncClient,
-    php_api: httpx.AsyncClient,
-) -> None:
-    py_response, php_response = await asyncio.gather(
-        py_api.get(f"/datasets/{dataset_id}"),
-        php_api.get(f"/data/{dataset_id}"),
-    )
-
-    if py_response.status_code == HTTPStatus.FORBIDDEN:
-        assert php_response.status_code == HTTPStatus.PRECONDITION_FAILED
-    else:
-        assert py_response.status_code == php_response.status_code
-
-    if py_response.status_code != HTTPStatus.OK:
-        # RFC 9457: Python API now returns problem+json format
-        assert py_response.headers["content-type"] == "application/problem+json"
-        # Both APIs should return error responses in the same cases
-        assert py_response.json()["code"] == php_response.json()["error"]["code"]
-        old_error_message = php_response.json()["error"]["message"]
-        assert py_response.json()["detail"].startswith(old_error_message)
-        return
-
-    try:
-        php_json = php_response.json()["data_set_description"]
-    except json.decoder.JSONDecodeError:
-        pytest.skip("A PHP error occurred on the test server.")
-
-    if "div" in php_json:
-        pytest.skip("A PHP error occurred on the test server.")
-
-    # There are a few changes between the old API and the new API, so we convert here:
-    # The new API has normalized `format` field:
-    php_json["format"] = php_json["format"].lower()
-
-    # Pydantic HttpURL serialization omits port 80 for HTTP urls.
-    php_json["url"] = php_json["url"].replace(":80", "")
-
-    # There is odd behavior in the live server that I don't want to recreate:
-    # when the creator is a list of csv names, it can either be a str or a list
-    # depending on whether the names are quoted. E.g.:
-    # '"Alice", "Bob"' -> ["Alice", "Bob"]
-    # 'Alice, Bob' -> 'Alice, Bob'
-    if (
-        "creator" in php_json
-        and isinstance(php_json["creator"], str)
-        and len(php_json["creator"].split(",")) > 1
-    ):
-        php_json["creator"] = [name.strip() for name in php_json["creator"].split(",")]
-
-    py_json = py_response.json()
-    if processing_data := py_json.get("processing_date"):
-        py_json["processing_date"] = str(processing_data).replace("T", " ")
-
-    manual = []
-    # ref test.openml.org/d/33 (contributor) and d/34 (creator)
-    #   contributor/creator in database is '""'
-    #   json content is []
-    for field in ["contributor", "creator"]:
-        if py_json[field] == [""]:
-            py_json[field] = []
-            manual.append(field)
-
-    if isinstance(py_json["original_data_url"], list):
-        py_json["original_data_url"] = ", ".join(str(url) for url in py_json["original_data_url"])
-
-    for field, value in list(py_json.items()):
-        if field in manual:
-            continue
-        if isinstance(value, int):
-            py_json[field] = str(value)
-        elif isinstance(value, list) and len(value) == 1:
-            py_json[field] = str(value[0])
-        if not py_json[field]:
-            del py_json[field]
-
-    if "description" not in py_json:
-        py_json["description"] = []
-
-    assert py_json == php_json
-
-
-@pytest.mark.parametrize(
-    "dataset_id",
-    [-1, 138, 100_000],
-)
-async def test_error_unknown_dataset(
-    dataset_id: int,
-    py_api: httpx.AsyncClient,
-) -> None:
-    response = await py_api.get(f"/datasets/{dataset_id}")
-
-    # The new API has "404 Not Found" instead of "412 PRECONDITION_FAILED"
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    # RFC 9457: Python API now returns problem+json format
-    assert response.headers["content-type"] == "application/problem+json"
-    error = response.json()
-    assert error["code"] == "111"
-    # instead of 'Unknown dataset'
-    assert error["detail"].startswith("No dataset")
-
-
-async def test_private_dataset_no_user_no_access(
-    py_api: httpx.AsyncClient,
-) -> None:
-    response = await py_api.get("/datasets/130")
-
-    # New response is 403: Forbidden instead of 412: PRECONDITION FAILED
-    assert response.status_code == HTTPStatus.FORBIDDEN
-    assert response.headers["content-type"] == "application/problem+json"
-    error = response.json()
-    assert error["code"] == "112"
-    assert error["detail"].startswith("No access granted")
-
-
-@pytest.mark.parametrize(
-    "api_key",
-    [ApiKey.DATASET_130_OWNER, ApiKey.ADMIN],
-)
-async def test_private_dataset_owner_access(
-    py_api: httpx.AsyncClient,
-    php_api: httpx.AsyncClient,
-    api_key: str,
-) -> None:
-    [private_dataset] = tests.constants.PRIVATE_DATASET_ID
-    py_response, php_response = await asyncio.gather(
-        py_api.get(f"/datasets/{private_dataset}?api_key={api_key}"),
-        php_api.get(f"/data/{private_dataset}?api_key={api_key}"),
-    )
-    assert php_response.status_code == HTTPStatus.OK
-    assert py_response.status_code == php_response.status_code
-    assert py_response.json()["id"] == private_dataset
 
 
 @pytest.mark.mut
